Move GameSeed class to rng module since it makes mor sense there
